feat: compute Var terms
