+++ pkg/builtin/default_function.go
@@ -1,4 +1,4 @@
-package pkg
+package builtin
 
 type DefaultFunction uint8
