+++ pkg/machine/value.go
@@ -1,0 +1,3 @@
+package machine
+
+type Value interface{}
