+++ pkg/syn/program.go
@@ -1,4 +1,4 @@
-package pkg
+package syn
 
 type Program[T Binder] struct {
 	version [3]uint32
